% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_vocabulary.R
\name{altitude_units_vocabulary}
\alias{altitude_units_vocabulary}
\title{Controlled vocabulary for the ALTITUDE_UNIT column, Sites table}
\usage{
altitude_units_vocabulary()
}
\value{
A character vector of altitude unit options
}
\description{
Returns controlled vocabulary options for altitude measurement units.
}
\examples{
altitude_units_vocabulary()
}
\seealso{
Other site: 
\code{\link{areas_vocabulary}()},
\code{\link{coordinate_systems_vocabulary}()},
\code{\link{countries_vocabulary}()},
\code{\link{example_sites_tibble}()},
\code{\link{geographic_features_sub_vocabulary}()},
\code{\link{geographic_features_vocabulary}()},
\code{\link{initialise_sites_tibble}()},
\code{\link{ocean_vocabulary}()}
}
\concept{site}
