% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{analytical_protocols_vocabulary}
\alias{analytical_protocols_vocabulary}
\title{Analytical Protocol Options Vocabulary}
\usage{
analytical_protocols_vocabulary()
}
\value{
A tibble with analytical protocol options
}
\description{
Returns analytical protocol options as a tibble with Protocol_Type, Short_Name, and Long_Name columns.
}
