% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_vocabulary.R
\name{areas_vocabulary}
\alias{areas_vocabulary}
\title{Ocean controlled vocabulary (wrapper for ocean_vocabulary())}
\usage{
areas_vocabulary()
}
\value{
A character vector of ocean name options
}
\description{
Returns controlled vocabulary options for IHO ocean regions.
}
\details{
Wrapper for \code{ocean_vocabulary()} used to maintain compatibility.
}
\examples{
head(areas_vocabulary(), 5)
}
\seealso{
Other site: 
\code{\link{altitude_units_vocabulary}()},
\code{\link{coordinate_systems_vocabulary}()},
\code{\link{countries_vocabulary}()},
\code{\link{example_sites_tibble}()},
\code{\link{geographic_features_sub_vocabulary}()},
\code{\link{geographic_features_vocabulary}()},
\code{\link{initialise_sites_tibble}()},
\code{\link{ocean_vocabulary}()}
}
\concept{site}
