% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_vocabulary.R
\name{countries_vocabulary}
\alias{countries_vocabulary}
\title{Countries controlled vocabulary}
\usage{
countries_vocabulary()
}
\value{
A character vector of country options
}
\description{
Returns controlled vocabulary options for country names based on ISO 3166-1 standard.
}
\details{
Provides country names from the \link[ISOcodes]{ISO_3166_1} standard, with additional
options for data quality flags.
}
\examples{
head(countries_vocabulary(), 5)
}
\seealso{
Other site: 
\code{\link{altitude_units_vocabulary}()},
\code{\link{areas_vocabulary}()},
\code{\link{coordinate_systems_vocabulary}()},
\code{\link{example_sites_tibble}()},
\code{\link{geographic_features_sub_vocabulary}()},
\code{\link{geographic_features_vocabulary}()},
\code{\link{initialise_sites_tibble}()},
\code{\link{ocean_vocabulary}()}
}
\concept{site}
