% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_dummy_data.R
\name{dummy_CREED_scores_tibble}
\alias{dummy_CREED_scores_tibble}
\title{Create Dummy CREED Scores Tibble}
\usage{
dummy_CREED_scores_tibble()
}
\value{
A tibble with 1 row of dummy CREED scores data
}
\description{
Creates a 1-row CREED scores summary tibble with test data.
Inherits column structure from initialise_CREED_scores_tibble().
}
