% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{dummy_parameters_vocabulary}
\alias{dummy_parameters_vocabulary}
\title{Dummy Parameters Data}
\usage{
dummy_parameters_vocabulary()
}
\value{
A data frame of dummy parameter data
}
\description{
Returns dummy parameter data combining quality parameters and chemical parameters.
}
