% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_vocabulary.R
\name{environ_compartments_vocabulary}
\alias{environ_compartments_vocabulary}
\title{Environmental compartments controlled vocabulary}
\usage{
environ_compartments_vocabulary()
}
\value{
A character vector of environmental compartment options
}
\description{
Returns controlled vocabulary options for broad environmental compartments.
Calls \code{names(environ_compartments_sub_vocabulary())} to return the first
level of the list.
}
\examples{
environ_compartments_vocabulary()
}
\seealso{
Other compartment: 
\code{\link{environ_compartments_sub_vocabulary}()},
\code{\link{example_compartments_tibble}()},
\code{\link{initialise_compartments_tibble}()},
\code{\link{measured_categories_vocabulary}()}
}
\concept{compartment}
