% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_example_data.R
\name{example_campaign_tibble}
\alias{example_campaign_tibble}
\title{Create Example Campaign Tibble}
\usage{
example_campaign_tibble()
}
\value{
A tibble with 1 row of dummy campaign data
}
\description{
Creates a 1-row campaign tibble with test data.
Inherits column structure from initialise_campaign_tibble().
}
