% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_example_data.R
\name{example_CREED_scores_tibble}
\alias{example_CREED_scores_tibble}
\title{Create Example CREED Scores Tibble}
\usage{
example_CREED_scores_tibble()
}
\value{
A tibble with 1 row of dummy CREED scores data
}
\description{
Creates a 1-row CREED scores summary tibble with test data.
Inherits column structure from \code{\link[=initialise_CREED_scores_tibble]{initialise_CREED_scores_tibble()}}.
}
\examples{
example_CREED_scores_tibble()

}
\seealso{
Other example_tibble: 
\code{\link{example_CREED_relevance_tibble}()},
\code{\link{example_CREED_reliability_tibble}()},
\code{\link{example_biota_tibble}()},
\code{\link{example_campaign_tibble}()},
\code{\link{example_compartments_tibble}()},
\code{\link{example_measurements_tibble}()},
\code{\link{example_methods_tibble}()},
\code{\link{example_parameters_tibble}()},
\code{\link{example_references_tibble}()},
\code{\link{example_samples_tibble}()},
\code{\link{example_sites_tibble}()}

Other CREED: 
\code{\link{CREED_choices_vocabulary}()},
\code{\link{CREED_tibble_to_mock_input}()},
\code{\link{example_CREED_relevance_tibble}()},
\code{\link{example_CREED_reliability_tibble}()},
\code{\link{initialise_CREED_data_tibble}()},
\code{\link{initialise_CREED_scores_tibble}()}
}
\concept{CREED}
\concept{example_tibble}
