% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_example_data.R
\name{example_measurements_tibble}
\alias{example_measurements_tibble}
\title{Create Example Measurements Tibble}
\usage{
example_measurements_tibble()
}
\value{
A tibble with 3 rows of dummy measurement data
}
\description{
Creates a 3-row measurements tibble with test data.
Inherits column structure from initialise_measurements_tibble().
}
