% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_example_data.R
\name{example_parameters_tibble}
\alias{example_parameters_tibble}
\title{Create Example Parameters Tibble}
\usage{
example_parameters_tibble()
}
\value{
A tibble with 3 rows of dummy parameter data
}
\description{
Creates a 3-row parameters tibble with test data.
Inherits column structure from \code{\link[=initialise_parameters_tibble]{initialise_parameters_tibble()}}.
}
\examples{
example_parameters_tibble()

}
\seealso{
Other example_tibble: 
\code{\link{example_CREED_relevance_tibble}()},
\code{\link{example_CREED_reliability_tibble}()},
\code{\link{example_CREED_scores_tibble}()},
\code{\link{example_biota_tibble}()},
\code{\link{example_campaign_tibble}()},
\code{\link{example_compartments_tibble}()},
\code{\link{example_measurements_tibble}()},
\code{\link{example_methods_tibble}()},
\code{\link{example_references_tibble}()},
\code{\link{example_samples_tibble}()},
\code{\link{example_sites_tibble}()}

Other parameter: 
\code{\link{initialise_parameters_tibble}()},
\code{\link{measured_types_vocabulary}()},
\code{\link{parameter_types_sub_vocabulary}()},
\code{\link{parameter_types_vocabulary}()},
\code{\link{parameters_vocabulary}()}
}
\concept{example_tibble}
\concept{parameter}
