% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_example_data.R
\name{example_sites_tibble}
\alias{example_sites_tibble}
\title{Create Example Sites Tibble}
\usage{
example_sites_tibble()
}
\value{
A tibble with 2 rows of dummy site data
}
\description{
Creates a 2-row sites tibble with test data.
Inherits column structure from initialise_sites_tibble().
}
