% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{fractionation_protocols_vocabulary}
\alias{fractionation_protocols_vocabulary}
\title{Fractionation Protocol Options Vocabulary}
\usage{
fractionation_protocols_vocabulary()
}
\value{
A tibble with fractionation protocol options
}
\description{
Returns fractionation protocol options as a tibble with Protocol_Type, Short_Name, and Long_Name columns.
}
