% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_vocabulary.R
\name{gender_vocabulary}
\alias{gender_vocabulary}
\title{Controlled vocabulary for the \code{SAMPLE_SPECIES_GENDER} column, Biota table}
\usage{
gender_vocabulary()
}
\value{
A character vector of gender options
}
\description{
Returns controlled vocabulary options for organism sex or gender.
}
\examples{
gender_vocabulary()
}
\seealso{
Other biota: 
\code{\link{example_biota_tibble}()},
\code{\link{initialise_biota_tibble}()},
\code{\link{lifestage_vocabulary}()},
\code{\link{species_groups_vocabulary}()},
\code{\link{species_names_vocabulary}()},
\code{\link{tissue_types_vocabulary}()}
}
\concept{biota}
