% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_vocabulary.R
\name{geographic_features_vocabulary}
\alias{geographic_features_vocabulary}
\title{Controlled vocabulary for the \code{SITE_GEOGRAPHIC_FEATURE} column, Sites table}
\usage{
geographic_features_vocabulary()
}
\value{
A character vector of geographic feature options
}
\description{
Returns controlled vocabulary options for site geographic features.
}
\examples{
geographic_features_vocabulary()
}
\seealso{
Other site: 
\code{\link{altitude_units_vocabulary}()},
\code{\link{areas_vocabulary}()},
\code{\link{coordinate_systems_vocabulary}()},
\code{\link{countries_vocabulary}()},
\code{\link{example_sites_tibble}()},
\code{\link{geographic_features_sub_vocabulary}()},
\code{\link{initialise_sites_tibble}()},
\code{\link{ocean_vocabulary}()}
}
\concept{site}
