% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{initialise_CREED_data_tibble}
\alias{initialise_CREED_data_tibble}
\title{Initialise CREED Scores Tibble}
\usage{
initialise_CREED_data_tibble()
}
\value{
A tibble with 0 rows and standardised CREED score columns
}
\description{
Creates an empty tibble with the standardised column structure for CREED
criterion scores. Used by both reliability and relevance modules to ensure
consistent data structure. This isn't a part of the externally-available
table structure, but we want to make sure it's harmonised locally anyway
}
