% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_vocabulary.R
\name{lifestage_vocabulary}
\alias{lifestage_vocabulary}
\title{Life stages controlled vocabulary}
\usage{
lifestage_vocabulary()
}
\value{
A character vector of life stage options
}
\description{
Returns controlled vocabulary options for organism life stages.
}
\examples{
head(lifestage_vocabulary(), 5)
}
\seealso{
Other biota: 
\code{\link{example_biota_tibble}()},
\code{\link{gender_vocabulary}()},
\code{\link{initialise_biota_tibble}()},
\code{\link{species_groups_vocabulary}()},
\code{\link{species_names_vocabulary}()},
\code{\link{tissue_types_vocabulary}()}
}
\concept{biota}
