% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_vocabulary.R
\name{measured_categories_vocabulary}
\alias{measured_categories_vocabulary}
\title{Controlled vocabulary for the \code{MEASURED_CATEGORY} column, Compartments table}
\usage{
measured_categories_vocabulary()
}
\value{
A named character vector of measured category options
}
\description{
Returns controlled vocabulary options for measurement context categories.
}
\details{
Provides standardised categories indicating where measurements were taken relative
to organisms (inside an organism, from its surface, or external to the organism (default))
}
\examples{
measured_categories_vocabulary()
}
\seealso{
Other compartment: 
\code{\link{environ_compartments_sub_vocabulary}()},
\code{\link{environ_compartments_vocabulary}()},
\code{\link{example_compartments_tibble}()},
\code{\link{initialise_compartments_tibble}()}

Other measurement: 
\code{\link{example_measurements_tibble}()},
\code{\link{initialise_measurements_tibble}()},
\code{\link{measured_flag_vocabulary}()},
\code{\link{measured_types_vocabulary}()},
\code{\link{parameter_unit_vocabulary}()},
\code{\link{uncertainty_types_vocabulary}()}
}
\concept{compartment}
\concept{measurement}
