% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_vocabulary.R
\name{measured_flag_vocabulary}
\alias{measured_flag_vocabulary}
\title{Controlled vocabulary for \code{MEASURED_FLAG}, Measurements table}
\usage{
measured_flag_vocabulary()
}
\value{
A character vector of measurement flag options
}
\description{
Returns controlled vocabulary options for measurement quality flags.
}
\details{
Provides standardised flags for indicating measurements below detection or quantification
limits. Options include:

\code{""} (empty string for measurements above limits)

\verb{< LOQ} (below limit of quantification)

\verb{< LOD} (below limit of detection)
}
\examples{
measured_flag_vocabulary()
}
\seealso{
Other measurement: 
\code{\link{example_measurements_tibble}()},
\code{\link{initialise_measurements_tibble}()},
\code{\link{measured_categories_vocabulary}()},
\code{\link{measured_types_vocabulary}()},
\code{\link{parameter_unit_vocabulary}()},
\code{\link{uncertainty_types_vocabulary}()}
}
\concept{measurement}
