% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{measured_flags_vocabulary}
\alias{measured_flags_vocabulary}
\title{Measurement Flags Controlled Vocabulary}
\usage{
measured_flags_vocabulary()
}
\value{
A character vector of measurement flag options
}
\description{
Returns measurement flag options.
}
