% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_vocabulary.R
\name{measured_types_vocabulary}
\alias{measured_types_vocabulary}
\title{Controlled vocabulary for the \code{MEASURED_TYPE} column, Parameters table}
\usage{
measured_types_vocabulary()
}
\value{
A character vector of measured type options
}
\description{
Returns controlled vocabulary options for the metric of pollution, quality, etc. measured.
}
\examples{
measured_types_vocabulary()
}
\seealso{
Other parameter: 
\code{\link{example_parameters_tibble}()},
\code{\link{initialise_parameters_tibble}()},
\code{\link{parameter_types_sub_vocabulary}()},
\code{\link{parameter_types_vocabulary}()},
\code{\link{parameters_vocabulary}()}

Other measurement: 
\code{\link{example_measurements_tibble}()},
\code{\link{initialise_measurements_tibble}()},
\code{\link{measured_categories_vocabulary}()},
\code{\link{measured_flag_vocabulary}()},
\code{\link{parameter_unit_vocabulary}()},
\code{\link{uncertainty_types_vocabulary}()}
}
\concept{measurement}
\concept{parameter}
