% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_vocabulary.R
\name{ocean_vocabulary}
\alias{ocean_vocabulary}
\title{Ocean controlled vocabulary}
\usage{
ocean_vocabulary()
}
\value{
A character vector of ocean name options
}
\description{
Returns controlled vocabulary options for IHO ocean regions.
}
\details{
Provides ocean and sea names from the International Hydrographic Organisation (IHO)
regions dataset, regions dataset (marineregions.org, World Seas IHO v3).
}
\examples{
head(ocean_vocabulary(), 5)
}
\seealso{
Other site: 
\code{\link{altitude_units_vocabulary}()},
\code{\link{areas_vocabulary}()},
\code{\link{coordinate_systems_vocabulary}()},
\code{\link{countries_vocabulary}()},
\code{\link{example_sites_tibble}()},
\code{\link{geographic_features_sub_vocabulary}()},
\code{\link{geographic_features_vocabulary}()},
\code{\link{initialise_sites_tibble}()}
}
\concept{site}
