% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_vocabulary.R
\name{parameter_types_sub_vocabulary}
\alias{parameter_types_sub_vocabulary}
\title{Controlled vocabulary for the \code{PARAMETER_TYPES_SUB} column, Parameters table}
\usage{
parameter_types_sub_vocabulary()
}
\value{
A character vector of parameter type subcategory options
}
\description{
Returns controlled vocabulary options for parameter type subcategories.
}
\details{
Dynamically built from unique \code{PARAMETER_TYPE_SUB} values in
\code{parameters_vocabulary()}, with "Mixture" appended and standard data quality
flags prepended. Chemical subtypes are taken from the
\href{http://classyfire.wishartlab.com}{ChemOnt} chemical taxonomy
}
\examples{
head(parameter_types_sub_vocabulary(), 15)
}
\seealso{
Other parameter: 
\code{\link{example_parameters_tibble}()},
\code{\link{initialise_parameters_tibble}()},
\code{\link{measured_types_vocabulary}()},
\code{\link{parameter_types_vocabulary}()},
\code{\link{parameters_vocabulary}()}
}
\concept{parameter}
