% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_vocabulary.R
\name{parameter_types_vocabulary}
\alias{parameter_types_vocabulary}
\title{Controlled vocabulary for the PARAMETER_TYPES column, Parameters table}
\usage{
parameter_types_vocabulary()
}
\value{
A character vector of parameter type options
}
\description{
Returns controlled vocabulary options for broad parameter classifications.
}
\examples{
parameter_types_vocabulary()
}
\seealso{
Other parameter: 
\code{\link{example_parameters_tibble}()},
\code{\link{initialise_parameters_tibble}()},
\code{\link{measured_types_vocabulary}()},
\code{\link{parameter_types_sub_vocabulary}()},
\code{\link{parameters_vocabulary}()}
}
\concept{parameter}
