% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{reference_type_vocabulary}
\alias{reference_type_vocabulary}
\title{Controlled vocabulary for reference types}
\usage{
reference_type_vocabulary()
}
\value{
A character vector of permitted reference types.
}
\description{
Returns the set of valid values for the \code{REFERENCE_TYPE} field in the
Reference table.
}
