% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_vocabulary.R
\name{species_groups_vocabulary}
\alias{species_groups_vocabulary}
\title{Controlled vocabulary for the SPECIES_GROUP column, Biota table}
\usage{
species_groups_vocabulary()
}
\value{
A character vector of species group options
}
\description{
Returns controlled vocabulary options for broad taxonomic groups from the EPA ECOTOX database.
}
\examples{
head(species_groups_vocabulary(), 5)
}
\seealso{
Other biota: 
\code{\link{example_biota_tibble}()},
\code{\link{gender_vocabulary}()},
\code{\link{initialise_biota_tibble}()},
\code{\link{lifestage_vocabulary}()},
\code{\link{species_names_vocabulary}()},
\code{\link{tissue_types_vocabulary}()}
}
\concept{biota}
