% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_vocabulary.R
\name{tissue_types_vocabulary}
\alias{tissue_types_vocabulary}
\title{Controlled vocabulary for the \code{SAMPLE_TISSUE} column, Biota table}
\usage{
tissue_types_vocabulary()
}
\value{
A character vector of tissue type options
}
\description{
Returns controlled vocabulary options for biological tissue types.
}
\details{
In future, this will be standardised with the BRENDA and/or Uberon tissue ontologies.
}
\examples{
tissue_types_vocabulary()
}
\seealso{
Other biota: 
\code{\link{example_biota_tibble}()},
\code{\link{gender_vocabulary}()},
\code{\link{initialise_biota_tibble}()},
\code{\link{lifestage_vocabulary}()},
\code{\link{species_groups_vocabulary}()},
\code{\link{species_names_vocabulary}()}
}
\concept{biota}
