% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_vocabulary.R
\name{uncertainty_types_vocabulary}
\alias{uncertainty_types_vocabulary}
\title{Uncertainty types controlled vocabulary}
\usage{
uncertainty_types_vocabulary()
}
\value{
A character vector of uncertainty type options
}
\description{
Returns controlled vocabulary options for statistical uncertainty and variability measures.
}
\examples{
head(uncertainty_types_vocabulary(), 5)
}
\seealso{
Other measurement: 
\code{\link{example_measurements_tibble}()},
\code{\link{initialise_measurements_tibble}()},
\code{\link{measured_categories_vocabulary}()},
\code{\link{measured_flag_vocabulary}()},
\code{\link{measured_types_vocabulary}()},
\code{\link{parameter_unit_vocabulary}()}
}
\concept{measurement}
