% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{altitude_units_vocabulary}
\alias{altitude_units_vocabulary}
\title{Controlled vocabulary for ALTITUDE_UNIT}
\usage{
altitude_units_vocabulary()
}
\value{
A character vector of altitude unit options
}
\description{
Returns controlled vocabulary options for altitude units.
}
