% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{areas_vocabulary}
\alias{areas_vocabulary}
\title{Areas Controlled Vocabulary}
\usage{
areas_vocabulary()
}
\value{
A character vector of area options
}
\description{
Returns controlled vocabulary options for areas (IHO ocean regions).
}
