% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_vocabulary.R
\name{coordinate_systems_vocabulary}
\alias{coordinate_systems_vocabulary}
\title{Coordinate systems controlled vocabulary}
\usage{
coordinate_systems_vocabulary(common_only = TRUE)
}
\arguments{
\item{common_only}{Return only the most common CRS options (default TRUE)}
}
\value{
A character vector of coordinate system options
}
\description{
Returns controlled vocabulary options for spatial coordinate reference systems.
}
\details{
Sourced from the EPSG Dataset via the \pkg{crsuggest} package.

More CRS will be added as needed. In the long term, vocabulary may be replaced
entirely for better standardisation with EPSG and/or PROJ4String.

CRS data sourced from \pkg{crsuggest} (Walker 2022).
Using the EPSG Dataset v10.019, a product of the International Association of
Oil & Gas Producers. Please view the terms of use at
\url{https://epsg.org/terms-of-use.html}.
}
\examples{
head(coordinate_systems_vocabulary(), 5)
}
\seealso{
Other site: 
\code{\link{altitude_units_vocabulary}()},
\code{\link{areas_vocabulary}()},
\code{\link{countries_vocabulary}()},
\code{\link{example_sites_tibble}()},
\code{\link{geographic_features_sub_vocabulary}()},
\code{\link{geographic_features_vocabulary}()},
\code{\link{initialise_sites_tibble}()},
\code{\link{ocean_vocabulary}()}
}
\concept{site}
