% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{coordinate_systems_vocabulary}
\alias{coordinate_systems_vocabulary}
\title{Coordinate Systems Controlled Vocabulary}
\usage{
coordinate_systems_vocabulary()
}
\value{
A character vector of coordinate system options
}
\description{
Returns controlled vocabulary options for coordinate systems.
}
