% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{countries_vocabulary}
\alias{countries_vocabulary}
\title{Countries Controlled Vocabulary}
\usage{
countries_vocabulary()
}
\value{
A character vector of country options
}
\description{
Returns controlled vocabulary options for countries.
}
