% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{CREED_choices_vocabulary}
\alias{CREED_choices_vocabulary}
\title{CREED Assessment Scoring Choices}
\usage{
CREED_choices_vocabulary()
}
\value{
Named character vector with CREED scoring choices
}
\description{
Returns the standardised CREED assessment scoring options
}
