% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_vocabulary.R
\name{data_source_vocabulary}
\alias{data_source_vocabulary}
\title{Controlled vocabulary for data sources}
\usage{
data_source_vocabulary()
}
\value{
A character vector of permitted data source types.
}
\description{
Returns the set of valid values for the \code{DATA_SOURCE} field in the
Reference table.
}
\examples{
data_source_vocabulary()
}
