% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{environ_compartments_sub_vocabulary}
\alias{environ_compartments_sub_vocabulary}
\title{Sub-compartment Options Mapping}
\usage{
environ_compartments_sub_vocabulary()
}
\value{
A named list of character vectors with sub-compartment options for each main compartment
}
\description{
Returns controlled vocabulary mapping for environmental sub-compartments organized by main compartment.
}
