% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{environ_compartments_vocabulary}
\alias{environ_compartments_vocabulary}
\title{Environmental Compartments Controlled Vocabulary}
\usage{
environ_compartments_vocabulary()
}
\value{
A character vector of environmental compartment options
}
\description{
Returns controlled vocabulary options for environmental compartments.
}
