% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{environ_compartments_vocabulary}
\alias{environ_compartments_vocabulary}
\title{Environmental compartments controlled vocabulary}
\usage{
environ_compartments_vocabulary()
}
\value{
A character vector of environmental compartment options
}
\description{
Returns controlled vocabulary options for broad environmental compartments.
}
\details{
Provides top-level categories for environmental matrices. Options include:

Aquatic

Atmospheric

Terrestrial

Biota

Not relevant

Not reported

Other
}
\seealso{
Other compartment: 
\code{\link{environ_compartments_sub_vocabulary}()},
\code{\link{initialise_compartments_tibble}()},
\code{\link{measured_categories_vocabulary}()}
}
\concept{compartment}
