% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_example_data.R
\name{example_biota_tibble}
\alias{example_biota_tibble}
\title{Create Example Biota Tibble}
\usage{
example_biota_tibble()
}
\value{
A tibble with 1 row of dummy biota data
}
\description{
Creates a 1-row biota tibble with test data.
Inherits column structure from initialise_biota_tibble().
}
