% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_example_data.R
\name{example_compartments_tibble}
\alias{example_compartments_tibble}
\title{Create Example Compartments Tibble}
\usage{
example_compartments_tibble()
}
\value{
A tibble with 2 rows of dummy compartment data
}
\description{
Creates a 2-row compartments tibble with test data.
Inherits column structure from initialise_compartments_tibble().
}
