% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_example_data.R
\name{example_measurements_tibble}
\alias{example_measurements_tibble}
\title{Create Example Measurements Tibble}
\usage{
example_measurements_tibble()
}
\value{
A tibble with 3 rows of dummy measurement data
}
\description{
Creates a 3-row measurements tibble with test data.
Inherits column structure from \code{\link[=initialise_measurements_tibble]{initialise_measurements_tibble()}}.
\code{SAMPLE_ID} values match those generated by \code{\link[=example_samples_tibble]{example_samples_tibble()}} and
\code{\link[=example_biota_tibble]{example_biota_tibble()}}. \code{REFERENCE_ID} matches \code{\link[=example_references_tibble]{example_references_tibble()}}.
}
\examples{
example_measurements_tibble()

}
\seealso{
Other example_tibble: 
\code{\link{example_CREED_relevance_tibble}()},
\code{\link{example_CREED_reliability_tibble}()},
\code{\link{example_CREED_scores_tibble}()},
\code{\link{example_biota_tibble}()},
\code{\link{example_campaign_tibble}()},
\code{\link{example_compartments_tibble}()},
\code{\link{example_methods_tibble}()},
\code{\link{example_parameters_tibble}()},
\code{\link{example_references_tibble}()},
\code{\link{example_samples_tibble}()},
\code{\link{example_sites_tibble}()}

Other measurement: 
\code{\link{initialise_measurements_tibble}()},
\code{\link{measured_categories_vocabulary}()},
\code{\link{measured_flag_vocabulary}()},
\code{\link{measured_types_vocabulary}()},
\code{\link{parameter_unit_vocabulary}()},
\code{\link{uncertainty_types_vocabulary}()}
}
\concept{example_tibble}
\concept{measurement}
