% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_example_data.R
\name{example_references_tibble}
\alias{example_references_tibble}
\title{Create Example References Tibble}
\usage{
example_references_tibble()
}
\value{
A tibble with 1 row of dummy reference data
}
\description{
Creates a 1-row references tibble with test data.
Inherits column structure from initialise_references_tibble().
REFERENCE_ID automatically generated using generate_reference_id().
}
