% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_example_data.R
\name{example_samples_tibble}
\alias{example_samples_tibble}
\title{Create Example Samples Tibble}
\usage{
example_samples_tibble()
}
\value{
A tibble with 2 rows of dummy sample data
}
\description{
Creates a 2-row samples tibble with test data.
Inherits column structure from initialise_samples_tibble().
}
