% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{extraction_protocols_vocabulary}
\alias{extraction_protocols_vocabulary}
\title{Controlled vocabulary for PROTOCOL_NAME and PROTOCOL_TYPE (extraction protocols)}
\usage{
extraction_protocols_vocabulary()
}
\value{
A tibble with extraction protocol options
}
\description{
Returns extraction protocol options as a tibble with Protocol_Type, Short_Name, and Long_Name columns.
}
