% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{gender_vocabulary}
\alias{gender_vocabulary}
\title{Initialise Gender Controlled Vocabulary}
\usage{
gender_vocabulary()
}
\value{
A character vector of gender options
}
\description{
Returns controlled vocabulary options for sample species gender.
}
