% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{gender_vocabulary}
\alias{gender_vocabulary}
\title{Gender controlled vocabulary}
\usage{
gender_vocabulary()
}
\value{
A character vector of gender options
}
\description{
Returns controlled vocabulary options for organism sex or gender.
}
\details{
Provides standardised categories for biological sex of sampled organisms. Options include:

Not reported

Not relevant

Male

Female

Mixed

Hermaphrodite

Other
}
\seealso{
Other biota: 
\code{\link{initialise_biota_tibble}()},
\code{\link{lifestage_vocabulary}()},
\code{\link{species_groups_vocabulary}()},
\code{\link{species_names_vocabulary}()},
\code{\link{tissue_types_vocabulary}()}
}
\concept{biota}
