% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_generate_IDs.R
\name{generate_reference_id}
\alias{generate_reference_id}
\title{Generate Reference ID}
\usage{
generate_reference_id(date, author, title)
}
\arguments{
\item{date}{Date (ACCESS_DATE or current date)}

\item{author}{Author string}

\item{title}{Title string}
}
\value{
Character string with format DateAuthorFirstThreeWords
}
\description{
Generate Reference ID
}
