% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{geographic_features_sub_vocabulary}
\alias{geographic_features_sub_vocabulary}
\title{Geographic Features Sub Controlled Vocabulary}
\usage{
geographic_features_sub_vocabulary()
}
\value{
A character vector of geographic feature subcategory options
}
\description{
Returns controlled vocabulary options for geographic feature subcategories.
}
