% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{geographic_features_vocabulary}
\alias{geographic_features_vocabulary}
\title{Geographic Features Controlled Vocabulary}
\usage{
geographic_features_vocabulary()
}
\value{
A character vector of geographic feature options
}
\description{
Returns controlled vocabulary options for geographic features.
}
