% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{initialise_CREED_scores_tibble}
\alias{initialise_CREED_scores_tibble}
\title{Initialise CREED Scores Tibble}
\usage{
initialise_CREED_scores_tibble()
}
\value{
A tibble with 0 rows and standardised CREED columns
}
\description{
Creates an empty tibble with the standardised column structure for CREED scores.
}
