% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{initialise_measurements_tibble}
\alias{initialise_measurements_tibble}
\title{Initialise Measurements Data Tibble}
\usage{
initialise_measurements_tibble()
}
\value{
A tibble with 0 rows and standardised measurement columns
}
\description{
Creates an empty tibble with the standardised column structure for measurements data.
}
