% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{lifestage_vocabulary}
\alias{lifestage_vocabulary}
\title{Initialise Life Stages Controlled Vocabulary}
\usage{
lifestage_vocabulary()
}
\value{
A character vector of life stage options
}
\description{
Returns controlled vocabulary options for sample species life stages.
}
