% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{measured_categories_vocabulary}
\alias{measured_categories_vocabulary}
\title{Measured Categories Controlled Vocabulary}
\usage{
measured_categories_vocabulary()
}
\value{
A character vector of measured category options
}
\description{
Returns controlled vocabulary options for measured categories.
}
