% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{measured_types_vocabulary}
\alias{measured_types_vocabulary}
\title{Measured Types Controlled Vocabulary}
\usage{
measured_types_vocabulary()
}
\value{
A character vector of measured type options
}
\description{
Returns controlled vocabulary options for measured types.
}
