% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{parameter_types_sub_vocabulary}
\alias{parameter_types_sub_vocabulary}
\title{Parameter Types Sub Controlled Vocabulary}
\usage{
parameter_types_sub_vocabulary()
}
\value{
A character vector of parameter type subcategory options
}
\description{
Returns controlled vocabulary options for parameter type subcategories.
}
