% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{parameter_types_vocabulary}
\alias{parameter_types_vocabulary}
\title{Parameter Types Controlled Vocabulary}
\usage{
parameter_types_vocabulary()
}
\value{
A character vector of parameter type options
}
\description{
Returns controlled vocabulary options for parameter types.
}
