% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{parameter_unit_vocabulary}
\alias{parameter_unit_vocabulary}
\title{Read in ecotoxicological units and conversion factors from csv}
\usage{
parameter_unit_vocabulary(select_column = NULL)
}
\arguments{
\item{select_column}{name of column to pull ("MEASURED_UNIT", "BASE_SI_UNIT", "CONVERSION_FACTOR", "UNIT_COMMENTS")}
}
\value{
a dataframe or a character vector
}
\description{
Read in ecotoxicological units and conversion factors from csv
}
