% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{protocol_categories_vocabulary}
\alias{protocol_categories_vocabulary}
\title{Protocol Categories Controlled Vocabulary}
\usage{
protocol_categories_vocabulary()
}
\value{
A character vector of protocol category options
}
\description{
Returns controlled vocabulary options for protocol categories.
}
