% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{protocol_options_vocabulary}
\alias{protocol_options_vocabulary}
\title{Protocol Options Data}
\usage{
protocol_options_vocabulary()
}
\value{
A tibble with Protocol_Type, Short_Name, and Long_Name columns for all protocols
}
\description{
Returns all protocol options data by combining all individual protocol vocabularies.
}
