% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{reference_character_limits}
\alias{reference_character_limits}
\title{reference_character_limits}
\usage{
reference_character_limits()
}
\value{
a list of character limits for fields in mod_references
}
\description{
reference_character_limits
}
