% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{sampling_protocols_vocabulary}
\alias{sampling_protocols_vocabulary}
\title{Sampling Protocol Options Vocabulary}
\usage{
sampling_protocols_vocabulary()
}
\value{
A tibble with sampling protocol options
}
\description{
Returns sampling protocol options as a tibble with Protocol_Type, Short_Name, and Long_Name columns.
}
