% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{species_groups_vocabulary}
\alias{species_groups_vocabulary}
\title{Species Groups Controlled Vocabulary}
\usage{
species_groups_vocabulary()
}
\value{
A character vector of species group options
}
\description{
Returns controlled vocabulary options for species groups. Taken from EPA ECOTOX db.
}
