% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{species_groups_vocabulary}
\alias{species_groups_vocabulary}
\title{Species groups controlled vocabulary}
\usage{
species_groups_vocabulary()
}
\value{
A character vector of species group options
}
\description{
Returns controlled vocabulary options for broad taxonomic groups from EPA ECOTOX database.
}
\details{
Provides high-level taxonomic groupings used in the ECOTOX database. Options include:

All

Algae

Amphibians

Bacteria

Birds

Crustaceans

Ecosystem

Fish

Fungi

Insects/Spiders

Invertebrates

Mammals

Molluscs

Moss/Hornworts

Plants

Reptiles

Worms

Other
}
\seealso{
Other biota: 
\code{\link{gender_vocabulary}()},
\code{\link{initialise_biota_tibble}()},
\code{\link{lifestage_vocabulary}()},
\code{\link{species_names_vocabulary}()},
\code{\link{tissue_types_vocabulary}()}
}
\concept{biota}
