% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{species_names_vocabulary}
\alias{species_names_vocabulary}
\title{Species Controlled Vocabulary}
\usage{
species_names_vocabulary()
}
\value{
A data frame of species information
}
\description{
Returns species name options from ECOTOX data
}
