% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{tissue_types_vocabulary}
\alias{tissue_types_vocabulary}
\title{Tissue types controlled vocabulary}
\usage{
tissue_types_vocabulary()
}
\value{
A character vector of tissue type options
}
\description{
Returns controlled vocabulary options for biological tissue types.
}
\details{
Provides standardised categories for biological tissues sampled from organisms.
Options include:

Not reported

Not relevant

Whole body

Total soft tissues

Muscle

Liver

Kidney

Fat/Adipose

Skin

Bone

Pyloric caeca

Body wall

Brain

Heart

Lung

Gill

Gonad

Shell

Carapace

Digestive Gland

Mantle

Blood

Egg

Larva

Leaf

Root

Stem

Fruit

Seed

Other
}
\seealso{
Other vocabulary: 
\code{\link{CREED_choices_vocabulary}()},
\code{\link{altitude_units_vocabulary}()},
\code{\link{analytical_protocols_vocabulary}()},
\code{\link{areas_vocabulary}()},
\code{\link{coordinate_systems_vocabulary}()},
\code{\link{countries_vocabulary}()},
\code{\link{dummy_parameters_vocabulary}()},
\code{\link{environ_compartments_sub_vocabulary}()},
\code{\link{environ_compartments_vocabulary}()},
\code{\link{extraction_protocols_vocabulary}()},
\code{\link{fractionation_protocols_vocabulary}()},
\code{\link{gender_vocabulary}()},
\code{\link{geographic_features_sub_vocabulary}()},
\code{\link{geographic_features_vocabulary}()},
\code{\link{lifestage_vocabulary}()},
\code{\link{measured_categories_vocabulary}()},
\code{\link{measured_flags_vocabulary}()},
\code{\link{measured_types_vocabulary}()},
\code{\link{parameter_types_sub_vocabulary}()},
\code{\link{parameter_types_vocabulary}()},
\code{\link{parameter_unit_vocabulary}()},
\code{\link{protocol_categories_vocabulary}()},
\code{\link{protocol_options_vocabulary}()},
\code{\link{sampling_protocols_vocabulary}()},
\code{\link{species_groups_vocabulary}()},
\code{\link{species_names_vocabulary}()},
\code{\link{uncertainty_types_vocabulary}()}

Other biota: 
\code{\link{gender_vocabulary}()},
\code{\link{initialise_biota_tibble}()},
\code{\link{lifestage_vocabulary}()},
\code{\link{species_groups_vocabulary}()},
\code{\link{species_names_vocabulary}()}
}
\concept{biota}
\concept{vocabulary}
