% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{tissue_types_vocabulary}
\alias{tissue_types_vocabulary}
\title{Initialise Tissue Types Controlled Vocabulary}
\usage{
tissue_types_vocabulary()
}
\value{
A character vector of tissue type options
}
\description{
Returns controlled vocabulary options for sample tissue types.
}
