% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_formats.R
\name{uncertainty_types_vocabulary}
\alias{uncertainty_types_vocabulary}
\title{Uncertainty Types Controlled Vocabulary}
\usage{
uncertainty_types_vocabulary()
}
\value{
A character vector of uncertainty type options
}
\description{
Returns controlled vocabulary options for uncertainty types commonly found
in scientific literature and databases.
}
